miniai/pyproject.toml
[tool.poetry]
name = "miniai"
version = "0.1.0"
description = "A small, benchmark-conscious agent orchestration framework"
authors = ["Elnamaki <Abdoelnamaki@gmail.com>"]
readme = "README"  # Optional: Add if you have a README
packages = [{include = "miniai"}]  # Specify the package directory

[tool.poetry.dependencies]
python = ">=3.11"
pydantic = ">=2.5.0"
httpx = ">=0.25.0"
aiohttp = ">=3.9.0"
tenacity = ">=8.2.0"
faiss-cpu = ">=1.7.4"
sentence-transformers = ">=2.2.2"
aiosqlite = ">=0.19.0"
aiocache = ">=0.12.2"
structlog = ">=23.2.0"
typer = ">=0.9.0"
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.0"
numpy = ">=1.24.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py311']